+++ src/interfaces/global/functions/IExpressHook.ts
