+++ src/interfaces/index.ts
@@ -3,12 +3,11 @@
 /* ---------------------------------------------------------------------------------- */
 export { default as IDynamicObject } from './IDynamicObject'
 export { default as IError } from './IError'
-export { default as IExpressHook } from './IExpressHook'
 export { default as IReject } from './IReject'
 export { default as IResolve } from './IResolve'
 
 /* ----------------------------------------------------------------------------------
- * Apollo Interfaces
+* Apollo Interfaces
 /* ---------------------------------------------------------------------------------- */
 export { default as IPaginatedData } from './apollo/IPaginatedData'
 // apollo > users
@@ -17,8 +16,10 @@ export { default as IUserCreateInput } from './apollo/users/IUserCreateInput'
 export { default as IUserUpdateInput } from './apollo/users/IUserUpdateInput'
 
 /* ----------------------------------------------------------------------------------
- * Global Directory Interfaces
+* Global Directory Interfaces
 /* ---------------------------------------------------------------------------------- */
+// global > functions
+export { default as IExpressHook } from './global/functions/IExpressHook'
 // global > utilities > activeRecords
 export { default as IActiveRecordMatchValues } from './global/utilities/activeRecords/IActiveRecordMatchValues'
 export { default as IActiveRecordSearchFilter } from './global/utilities/IActiveRecordSearchFilter'
